Move overlay files to directory, move all stories to features as there isn't a clear default.
