+++ docs/content/Overlay.mdx
@@ -7,7 +7,7 @@ source: https://github.com/primer/react/blob/main/src/Overlay.tsx
 storybook: '/react/storybook?path=/story/private-components-overlay--dropdown-overlay'
 ---
 
-import data from '../../src/Overlay.docs.json'
+import data from '../../src/Overlay/Overlay.docs.json'
 
 ```js
 import {Overlay} from '@primer/react'
