+++ src/Overlay/Overlay.tsx
@@ -1,15 +1,15 @@
 import styled from 'styled-components'
 import React, {ComponentPropsWithRef, ReactElement, useEffect, useRef} from 'react'
-import useLayoutEffect from './utils/useIsomorphicLayoutEffect'
-import {get} from './constants'
-import {AriaRole, Merge} from './utils/types'
-import {useOverlay, TouchOrMouseEvent} from './hooks'
-import Portal from './Portal'
-import sx, {SxProp} from './sx'
-import {useRefObjectAsForwardedRef} from './hooks/useRefObjectAsForwardedRef'
+import useLayoutEffect from '../utils/useIsomorphicLayoutEffect'
+import {get} from '../constants'
+import {AriaRole, Merge} from '../utils/types'
+import {useOverlay, TouchOrMouseEvent} from '../hooks'
+import Portal from '../Portal'
+import sx, {SxProp} from '../sx'
+import {useRefObjectAsForwardedRef} from '../hooks/useRefObjectAsForwardedRef'
 import type {AnchorSide} from '@primer/behaviors'
-import {useTheme} from './ThemeProvider'
-import {ForwardRefComponent as PolymorphicForwardRefComponent} from './utils/polymorphic'
+import {useTheme} from '../ThemeProvider'
+import {ForwardRefComponent as PolymorphicForwardRefComponent} from '../utils/polymorphic'
 
 type StyledOverlayProps = {
   width?: keyof typeof widthMap
