+++ src/Overlay/index.ts
@@ -1,0 +1,1 @@
+export {default, OverlayProps} from './Overlay'
